wip: Save state before backtracking
client_info.port is port on which we receive datagrams from the client.
Therefore, we can't use this client_info instance to create
a reserved worker for the new allocation,
as the client will connect from different port.
We have to rely on dispatcher to create the new worker process,
but at the same time store the reservation somewhere,
so that it can be picked up by that new worker when it reads
the RESERVATION-TOKEN attribute coming with an allocate request.